rename LicenseManagement to SubscriptionManagement
