+++ src/components/subscriptions/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as LicenseManagementPage } from './LicenseManagementPage';
+export { default as SubscriptionManagementPage } from './SubscriptionManagementPage';
